Reimplement the tick box as a state machine
For my own sanity, and simplicity of the conditions.  Now the implementation
can be checked mostly separately from the logic.

The UML State diagram was made in UMLet.

Also, fix the Good tactics so that they don't use evars.  Now the proof to
mbtaGood' is fully automated, modulo that pesky guard condition...