Add a server-syncing box
It's quite a monster at 520 lines. The description of the algorithm is 75 lines.
There's no particularly complicated logic (I ended up implementing a simpler
version of the protocol, where the ui/client/master wiring diagram is
responsible for scheduling the tick boxes so that the get-update finishes
shortly before the encryption starts, which itself finishes shortly before the
update should be pushed to the server.  (I'll probably need to go fix the tick
boxes to be pinned to intervals in terms of system ticks, rather than a number
of ticks after the previous push.)

Much of the pain seems to come from hard-wiring the boxes so that the tick boxes
and the encryption/decryption boxes are hidden from the master wiring diagram.

Perhaps there's a better way to do this: it might make sense to define a type of
internal events, and specify the appropriate bits of the wiring diagram for the
internal events, and leave it to the master wiring diagram to actually wire
things together.  Or perhaps we just need better syntax.